fix: refactor `read_file_from_workspace` command to gitbutler-repo
