+++ crates/gitbutler-repo/src/commands.rs
@@ -2,6 +2,7 @@ use anyhow::Result;
 use bstr::BString;
 use gitbutler_command_context::CommandContext;
 use gitbutler_project::Project;
+use std::path::Path;
 
 use crate::{Config, RepositoryExt};
 
@@ -11,6 +12,7 @@ pub trait RepoCommands {
     fn get_local_config(&self, key: &str) -> Result<Option<String>>;
     fn set_local_config(&self, key: &str, value: &str) -> Result<()>;
     fn check_signing_settings(&self) -> Result<bool>;
+    fn read_file_from_workspace(&self, relative_path: &Path) -> Result<String>;
 }
 
 impl RepoCommands for Project {
@@ -46,5 +48,27 @@ impl RepoCommands for Project {
         let ctx = CommandContext::open(self)?;
         ctx.repository().remote(name, url)?;
         Ok(())
+    }
+
+    fn read_file_from_workspace(&self, relative_path: &Path) -> Result<String> {
+        let ctx = CommandContext::open(self)?;
+        let base_path = ctx
+            .repository()
+            .path()
+            .parent()
+            .ok_or(anyhow::anyhow!("Base path is invalid"))?;
+
+        let canonicalized_file_path = base_path.join(relative_path).canonicalize()?;
+
+        if canonicalized_file_path.as_path().starts_with(base_path) {
+            let tree = ctx.repository().head()?.peel_to_tree()?;
+            let entry = tree.get_path(relative_path)?;
+            let blob = ctx.repository().find_blob(entry.id())?;
+            let content = std::str::from_utf8(blob.content())?;
+
+            Ok(content.to_string())
+        } else {
+            anyhow::bail!("Invalid file path");
+        }
     }
 }
